Implement console version
This is an implementation of the Durák card game.
Use the Main method to run it once and check the console output.
Only the AI plays in this version. (4 players by default.)
Use the StabilityTest to run the game 100 times and assert that no errors are thrown, meaning the game logic is correct.